# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MovieLand",
    "version": "1.0.0",
    "description": "This is the API documentation for the Movieland application."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/api/v1/movies": {
      "get": {
        "summary": "Retrieve all movies",
        "description": "Get a list of all movies.",
        "responses": {
          "200": {
            "description": "A list of movies."
          }
        }
      },
      "post": {
        "summary": "Create a new movie",
        "description": "Add a new movie to the database.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "genre": {
                    "type": "string"
                  },
                  "releaseDate": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Movie created successfully."
          }
        }
      }
    },
    "/api/v1/movies/{id}": {
      "get": {
        "summary": "Retrieve a movie by ID",
        "description": "Get a movie by its unique ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the movie to retrieve.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Movie data retrieved successfully."
          },
          "404": {
            "description": "Movie not found."
          }
        }
      },
      "put": {
        "summary": "Update an existing movie",
        "description": "Modify the details of an existing movie.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the movie to update.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "genre": {
                    "type": "string"
                  },
                  "releaseDate": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Movie updated successfully."
          },
          "404": {
            "description": "Movie not found."
          }
        }
      },
      "delete": {
        "summary": "Delete a movie",
        "description": "Remove a movie from the database.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the movie to delete.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Movie deleted successfully."
          },
          "404": {
            "description": "Movie not found."
          }
        }
      }
    }
  },
  "tags": []
}